fix(cmd/build): Improve push (<<) op runtime fn for JS native strings
Previously, the --web runtime's builtin function for the push (<<)
operator in Oak assumed that all inputs were either strings promoted to
Oak runtime strings (__Oak_String) or Oak lists (JS arrays). When a JS
native string was passed to it by e.g. directly accessing a
string-valued property of a JS-native object, this would cause the
operator function to crash the program.

This is now fixed by removing the incorrect assumption, and making the
operator builtin function auto-promote every input to an __Oak_String if
it is a JS native string.