fix(lib/syntax): libsyntax positions [line, col] -> [offset, line, col]
For most operations on syntax constructs, a byte offset into the program
source text is much more useful than a (line, col) pair. The (line, col)
position is only really useful for when token positions need to be
directly human-readable, like in error message outputs. By contrast, for
most machine operations like syntax highlighting, pointing to the error
source location in an error message, and so on, a file offset is much
more useful.

This breaking API change introduces a new component of token positions
into the libsyntax token object, which includes a byte offset into the
source text of a parsed program. It begins at 0, and goes to a max value
of len(source). All libsyntax tests and cmd/build, the primary dependent
of libsyntax, have been updated to match this new token position type.

This change anticipates syntax highlighting support in `oak cat`.