feat(eval): Short-circuit boolean operators when possible
This patch modifies the behavior of and (&) and or (|) binary operators
in Oak so that if the left operand is a boolean value and the expression
can short-circuit, it does. This behavior is a backwards-compatible
change for all code with no side effects in boolean expression operands
(which should be all reasonable code) and applies to all runtimes.

This change, despite introducing some logic, has a net performance win
because it lets the runtime skip past unnecessary work. The runtime
performance improvements on the standard library tests are:

        2% on the native (Go) interpreter
        3% on the web (oak build --web) build